YR-990 [接口]写手个人主页-我的文章（文章发布：视屏/图文）

+++ writer-article/writer-article-api/src/main/java/com/yryz/writer/modules/draft/DraftApi.java
@@ -47,7 +47,7 @@ public interface DraftApi {
      * @param draft
      * @return
      */
-    RpcResponse<Integer> add(Draft draft);
+    RpcResponse<Long> add(Draft draft);
 
     RpcResponse<Integer> check(Draft draft);
